==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/01_scaleu_w210_satellite_data_scale_up.ipynb <==
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

from shapely.geometry import Point, Polygon
import geopandas as gpd

# my script
from w210_attribute_library_scale import tilekey, haversine_distance, withinstates, findcounty, findcountyfp

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

## Read Satellite Attributes

satd = pd.read_csv(attrs+"scaleup_fl_CNN_predictions_07_05_2022.csv")
satd.head(2)

# Select a Few to Test Scripts First
satt = satd.sample(frac=0.001)
len(satt)

satt = satd
len(satt)

### Assign County and Create Key

# Read shapefile with the county polygons, and select Florida counties in flcounty
datsubdir = "../data/shapefile/"
shapedir = 'cb_2018_us_county_500k/'
shapefile500 = "cb_2018_us_county_500k.shp"

us_county500 = gpd.read_file(datsubdir+shapedir+shapefile500)

flcounty = us_county500[(us_county500["STATEFP"]=='12')]
print(len(flcounty))

# Create Dictionary for Florida Counties - FP:NAME
flcounty_dict = {}
for index, row in flcounty.iterrows():
    flcounty_dict[row["COUNTYFP"]] = row["NAME"]
    
flcounty_dict["No_Florida"] = "No_Florida"

satt["county_fp"] = satt.apply(lambda row: findcountyfp(Point(row["lon"],row["lat"]), flcounty), axis=1)
satt["Key"] = satt.apply(lambda row: str(row["county_fp"])+"_"+str(row.name), axis=1)
satt["County"] = satt.apply(lambda row: flcounty_dict[row["county_fp"]], axis=1)

nf = satt[satt["county_fp"]=="No_Florida"]
len(nf)

satt.to_csv(attrs+"scale_up_satellite_key_county.csv")

sattfl = satt[satt["county_fp"] != "No_Florida"]
len(sattfl)

sattfl.to_csv(attrs+"scale_up_F_satellite_key_county.csv")
