==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/1_scaleu_w210_match_ws_to_tile_florida.ipynb <==
# To Match WS to Tile
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

# my script
from w210_attribute_library_scale import tilekey, haversine_distance, date_string

print("Date of reference: ", date_string)

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

fweather = 'weather_data.csv'
dfw = pd.read_csv(datdir+fweather)
print(len(dfw))
dfw.head(1)

### Check and Select Working/Valid Weather Stations

wskey = dfw["usaf_wban"].unique()

dated = datetime.strptime(date_string,"%Y/%m/%d")

dfw["DateD"] = dfw.apply(lambda row: datetime.strptime(str(row["year"])+"-"+str(row["mo"])+"-"+str(row["da"]),"%Y-%m-%d") ,axis=1)

i = 0
wsgood = []
wsnogood = []
for wsname in wskey:
    td = timedelta(365*3)
    dftemp = dfw[(dfw["usaf_wban"] == wsname)]
    if (dftemp["DateD"].max() > dated) & (dftemp["DateD"].min() < dated - td):
        wsgood.append(wsname)
    else:
        wsnogood.append(wsname)

wsnogood_set = set(wsnogood)
wsnogood_list = list(wsnogood_set)
print(wsnogood_list)

dfw = dfw[~dfw["usaf_wban"].isin(wsnogood_list)]
print(len(wskey), len(dfw["usaf_wban"].unique()))

ws = dfw[['stn_wban', "name", "lon", "lat"]].drop_duplicates()
ws.columns = ["Key", 'name', 'lon_w', 'lat_w']
print(len(ws))

**For Reference - list of no-good weather stations**  
`
['722261_99999', '999999_53848', '998199_99999', '747940_99999', '722118_99999', '693254_99999', '720374_92825', '998305_99999', '690524_99999', '722012_92817', '722224_53862', '749048_99999', '999999_63890', '997733_99999', '747761_99999']
`

duplicateName = ['WHITING FIELD NAVAL AIR STATI', 'KEYSTONE AIRPARK']
ws[(ws['name'].isin(duplicateName))]

# Replace values with NaN
no_valid = [99.99000, 999.900000, 9999.900000, 99999.000000]
ws = ws.replace(99.99000,np.nan)
ws = ws.replace(999.9000000,np.nan)
ws = ws.replace(9999.900000,np.nan)
ws = ws.replace(99999.000000,np.nan)

## Read Tile Data

ftileslarge = 'scale_up_F_satellite_key_county.csv'
dftiles_large = pd.read_csv(attrs+ftileslarge)
dftiles_large.head(1)

## Merge Tile and Weather for Distance Calculation

`result = pd.merge(left, right, how="left", on=["key1", "key2"])`

**Reference:**
https://pandas.pydata.org/docs/user_guide/merging.html

dflarge = pd.merge(dftiles_large,ws, how="cross")
print(len(dflarge))
dflarge.head(1)

## Calculate The Distance

dflarge['Distance'] = dflarge.apply(lambda row: 
                                    haversine_distance(row['lat'], row['lon'], 
                                                       row['lat_w'], row['lon_w'], 
                                                       earth_radius=3963.19), axis=1)
dflarge.head(2)

## Find the Minimum Distances within Tile and Weather Station

`df.groupby('Company')['MPG'].agg('min')`  
`df.groupby('Company')[['MPG', 'EngineSize']].agg('min')`

**Reference:**
https://datascienceparichay.com/article/pandas-groupby-minimum/

# Find the rows with the minimum Distance for each Key_x
dfmin1 = dflarge.groupby(['Key_x'])['Distance'].min().to_frame()
print(len(dfmin1))

# Select only the rows with the minimum
keysL = list(dfmin1.index)
minD = list(dfmin1['Distance'])
dfF1 = dflarge[((dflarge['Key_x'].isin(keysL)) &  (dflarge['Distance'].isin(minD)))]

**Reference Duplicates:**  
https://sparkbyexamples.com/pandas/pandas-get-list-of-all-duplicate-rows/#:~:text=Pandas%20DataFrame.,multiple%20columns%20or%20all%20columns.

#Checking for Duplicates
df2 = dfF1[dfF1["Key_x"].duplicated()==True]
dup1 = df2["Key_x"].unique()
# dup1 = ['1082_0_1','1083_0_1', '2406_0_1', '2459_0_1', '2463_0_1', '2737_0_1', '3294_0_2', '3294_1_0', '556_0_1']
dfF1[(dfF1['Key_x'].isin(dup1))]

dfF1.drop_duplicates(subset=['Key_x'], inplace=True)
len(dfF1)

# dfF1["Distance"].describe()

dfF1.to_csv(datdir+"scale_up_F_tiles_ws_pair.csv", index=False)

dfF1.head()

## Get Pair for Small Tile

### Read Small Tile Data

ftilesmall = 'smalltiles.csv'
dftiles_small = pd.read_csv(datdir+ftilesmall)
dftiles_small["Key"] = dftiles_small.apply(lambda row: tilekey(row), axis=1)
dftiles_small = dftiles_small[["Key", "lon", "lat", "County", "CountyFP"]]
dftiles_small.columns = ["Key", "lon_t", "lat_t", "County", "CountyFp"]

print(len(dftiles_small))
dftiles_small.head(2)

### Merge Tile and Weather for Distance Calculation

dfsmall = pd.merge(dftiles_small,ws, how="cross")
print(len(dfsmall))
dfsmall.head(1)

### Calculate The Distance

dfsmall['Distance'] = dfsmall.apply(lambda row: 
                                    haversine_distance(row['lat_t'], row['lon_t'], 
                                                       row['lat_w'], row['lon_w'], 
                                                       earth_radius=3963.19), axis=1)

### Find the Minimum Distances within Tile and Weather Station

dfminSmall = dfsmall.groupby(['Key_x'])['Distance'].min().to_frame()

keysL = list(dfminSmall.index)
minD = list(dfminSmall['Distance'])
dfF2 = dfsmall[((dfsmall['Key_x'].isin(keysL)) &  (dfsmall['Distance'].isin(minD)))]
print(len(dfF2))
dfF2.head(2)

#Checking for Duplicates
df2 = dfF2[dfF2["Key_x"].duplicated()==True]
dup1 = dfF2["Key_x"].unique()
# dup1 = ['1082_0_1','1083_0_1', '2406_0_1', '2459_0_1', '2463_0_1', '2737_0_1', '3294_0_2', '3294_1_0', '556_0_1']
dfF2[(dfF2['Key_x'].isin(dup1))]

dfF2.drop_duplicates(subset=['Key_x'], inplace=True)
len(dfF2)

dfF2["Distance"].describe()

dfF2.to_csv(datdir+"scale_up_small_tile_ws_pair.csv", index=False)

len(dfF2["Key_y"].unique())
