==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/02_scaleU_karst_data.ipynb <==
# !pip install shapely

import pandas as pd
import numpy as np
from zipfile import ZipFile
import os
from os.path import basename
import urllib
import shutil
from geopandas import gpd
from shapely.geometry import Point, Polygon
# import shapefile

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

gdf = gpd.read_file("../data/sink_density_classified_polys_1km_conus/sink_density_classified_polys_1km_conus.shp")
# geo_df = GeoDataFrame(df, crs=crs, geometry=geometry)

new_gdf = gdf.to_crs(epsg=4326)

new_gdf.head()

new_gdf['x_coord'] = new_gdf.apply(lambda row: row['geometry'].centroid.x, axis=1)
new_gdf['y_coord'] = new_gdf.apply(lambda row: row['geometry'].centroid.y, axis = 1)

new_gdf.to_csv(datdir+"karst_gridcode_raw_data.csv")
