==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/2_scaleu_w210_get_ws_attr_florida.ipynb <==
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

# my script
from w210_attribute_library_scale import tilekey, haversine_distance, getwf_final, weatherFeaturesSU, date_string

print("Date of reference: ", date_string)

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

# IMPORTANT - SET THE DATA OF REFERENCE
# date_string = '2022/06/01'

fweather = 'weather_data.csv'
dfw = pd.read_csv(datdir+fweather)
print(len(dfw))
dfw.head(1)

colweather = ['stn_wban', 'year', 'mo', 'da', 'temp', 'prcp', 'flag_prcp', 'min', 'max',
       'flag_prcp', 'rain_drizzle', 'sndp', 'fog', 'snow_ice_pellets', 'hail',
       'thunder', 'tornado_funnel_cloud', 'wdsp', 'mxpsd', 'gust', 'visib',
       'dewp', 'slp', 'usaf_wban', 'wban']
dfw1 = dfw[colweather]
dfw1.head(1)

# Create a Date Field to be able to make date comparisons
dfw1["DateD"] = dfw1.apply(lambda row: datetime.strptime(str(row["year"])+"-"+str(row["mo"])+"-"+str(row["da"]),"%Y-%m-%d") ,axis=1)
dfw1.head(2)

# Replace values with NaN
no_valid = [99.99000, 999.900000, 9999.900000, 99999.000000]
dfw1 = dfw1.replace(99.99000,np.nan)
dfw1 = dfw1.replace(999.9000000,np.nan)
dfw1 = dfw1.replace(9999.900000,np.nan)
dfw1 = dfw1.replace(99999.000000,np.nan)

dfw1["DateD"].max(), dfw1["DateD"].min()

### Get Weather Features for Large Tiles

#### Using Timedelta=0 

**Note:** For model application, we only need timedelta zero

# Read file with large tiles
dfF2 = pd.read_csv(datdir+ "scale_up_F_tiles_ws_pair.csv")
dfF2.head(1)

daysdelta = timedelta(0)
is_scale = True
fields = ["Key_x", "Key_y", "name_y", "lon", "lat", "lon_w", "lat_w", "Distance"]
dfwf_large_0 = weatherFeaturesSU(dfF2, dfw1, is_scale, daysdelta, date_string, fields)
dfwf_large_0.head(1)

dfwf_large_0.to_csv(attrs + "scale_up_F_tiles_ws_attr.csv", index=False)
