==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/3_scaleu_w210_soil_attributes.ipynb <==
# To Match Karst to Tile
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

from shapely.geometry import Point, Polygon
import geopandas as gpd

# my script
from w210_attribute_library_scale import tilekey, haversine_distance, withinstates

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

## Read Soil Data

soild = pd.read_csv(datdir+'FL_soil_byCounty.csv')
print(len(soild))

soild.columns

## Read Tiles

ftileslarge = 'scale_up_F_satellite_ws_attr.csv'
dftiles_large = pd.read_csv(attrs+ftileslarge)
# dftiles_large["Key"] = dftiles_large.apply(lambda row: tilekey(row), axis=1)
# dftiles_large = dftiles_large[["Key", "lon", "lat", "County", "CountyFP"]]
# dftiles_large.columns = ["Key", "lon_t", "lat_t", "County", "CountyFp"]

print(len(dftiles_large))
dftiles_large.head(2)

# Cross Merge Dataframes
dffinal = pd.merge(dftiles_large,soild, how="cross")
print(len(dffinal))

dffinal.columns

# Get Distance
dffinal['Distance'] = dffinal.apply(lambda row: 
                                    haversine_distance(row['lat_t'], row['lon_t'], 
                                                       row['Center Latitude'], 
                                                       row['Center Longitude'], 
                                                       earth_radius=3963.19), axis=1)

## Find the Minimum Distances within Tile and Soil Point

`df.groupby('Company')['MPG'].agg('min')`  
`df.groupby('Company')[['MPG', 'EngineSize']].agg('min')`

**Reference:**
https://datascienceparichay.com/article/pandas-groupby-minimum/

dfmin1 = dffinal.groupby(['Key'])['Distance'].min().to_frame()

keysL = list(dfmin1.index)
minD = list(dfmin1['Distance'])
dfF1 = dffinal[((dffinal['Key'].isin(keysL)) &  (dffinal['Distance'].isin(minD)))]
len(dfF1)

dfF1.to_csv(attrs+"scale_up_F_satellite_ws_soil_attr.csv", index=False)
