==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/4_scaleu_w210_karst_data.ipynb <==
# To Match Karst to Tile
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

from shapely.geometry import Point, Polygon
import geopandas as gpd

# my script
from w210_attribute_library_scale import tilekey, haversine_distance, withinstates

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

### Read Karst Data

fkarst = 'karst_gridcode_raw_data.csv'
dfk = pd.read_csv(datdir+fkarst)
print(len(dfk))
dfk.head(1)

#### Select Relevant Points

subdir = "../data/shapefile/"
shapedir = 'cb_2018_us_state_500k/'
shapefile500 = "cb_2018_us_state_500k.shp"

us500 = gpd.read_file(subdir+shapedir+shapefile500)

flgeometry = list(us500[(us500["NAME"]=='Florida')]["geometry"])[0]
gageometry = list(us500[(us500["NAME"]=='Georgia')]["geometry"])[0]
algeometry = list(us500[(us500["NAME"]=='Alabama')]["geometry"])[0]

geometries = [flgeometry, gageometry, algeometry ]
# geometries = [flgeometry]

# dfk["Florida"] = dfk.apply(lambda row: "FL" if (Point(row["x_coord"],row["y_coord"]).within(flgeometry)) else "NoFL", axis=1)

dfk["in_relevant_state"] = dfk.apply(lambda row: withinstates(geometries, Point(row["x_coord"],row["y_coord"])), axis=1)

dfk = dfk[dfk["in_relevant_state"] == "Yes"]
len(dfk)

### Read Current Tile Data

ftileslarge = 'scale_up_F_satellite_ws_soil_attr.csv'
dftiles_large = pd.read_csv(attrs+ftileslarge)
dftiles_large.head(1)

dftiles_large.columns

dftiles_large = dftiles_large[['Key', 'Calcium Carbonate', 'Gypsum',
       'Soil Health Organic Matter', 'Percent Clay', 'Percent Sand',
       'Percent Silt', 'Available Water Storage', 'date_ws', 'name_x', 'lon_t',
       'lat_t', 'rolling_7_precip', 'rolling_15_precip', 'rolling_30_precip',
       'rolling_60_precip', 'rolling_90_precip', 'y1_mean_prc', 'y1_max_prc',
       'y1_mean_tmp', 'y1_max_tmp', 'y1_min_tmp', 'y2_mean_prc', 'y2_max_prc',
       'y2_mean_tmp', 'y2_max_tmp', 'y2_min_tmp', 'y3_mean_prc', 'y3_max_prc',
       'y3_mean_tmp', 'y3_max_tmp', 'y3_min_tmp', 'County', 'county_fp',
       'name_y', 'imagenum', 'geometry', 'AnnualCrop', 'Forest',
       'HerbaceousVegetation', 'Highway', 'Industrial', 'Pasture',
       'PermanentCrop', 'Residential', 'River', 'SeaLake', 'prediction',
       'prediction_name']]

# dft = dftiles_large.sample(frac=0.001)
# len(dft)

def findkarst(tile, dfk):
    
    i = 0
    
    for index, rkarst in dfk.iterrows():
        d = haversine_distance(tile['lat_t'], tile['lon_t'], rkarst["y_coord"], rkarst["x_coord"])
        if i == 0:
            minv = d
            kdata = rkarst["gridcode"]
            i += 1
        else:
            if d < minv:
                minv = d
                kdata = rkarst["gridcode"]
    
    return kdata

dftiles_large['gridcode'] = dftiles_large.apply(lambda row: findkarst(row, dfk), axis=1)

dftiles_large.to_csv("scale_up_F_satellite_ws_soil_karst_attr.csv", index=False)

dftiles_large.to_csv(attrs+"scale_up_F_satellite_ws_soil_karst_attr.csv", index=False)

dftiles_large["gridcode"].unique()

dftiles_large.head(1)

dftiles_large.columns
