==> Sinkhole-Intelligence-Capstone/08_Scale_Up_Process/1_Scale_Up_Generate_Attributes/5_scaleu_w210_sinkhole_features.ipynb <==
import fiona
from shapely.geometry import Point, Polygon
import geopandas as gpd
import pandas as pd

import numpy as np
import pandas as pd
import datetime
from datetime import datetime
import os
import datetime
from datetime import datetime, timedelta
import matplotlib.pyplot as plt
import math

from w210_attribute_library_scale import shAttributes, getdate, haversine_distance, tilekey, date_string

print("Date of reference: ", date_string)

datdir = '../data/'
attrs = "../attrs/"

## Read Sinkhole Data

df_sh = pd.read_excel(datdir+"Florida_Subsidence_Incident_Reports.xlsx", sheet_name=0)
df_sh["DateD"] = df_sh.apply(lambda row: getdate(row["EVENT_DATE"]), axis=1)
df_sh = df_sh[~df_sh["EVENT_DATE"].isna()]
print(len(df_sh))
df_sh.head(2)

## Read Tile Data

ftileslarge = 'scale_up_F_satellite_ws_soil_karst_attr.csv'
dftiles_large = pd.read_csv(attrs+ftileslarge)
# dftiles_large["Key"] = dftiles_large.apply(lambda row: tilekey(row), axis=1)
dftiles_large["DateD"] = datetime.strptime(date_string,"%Y/%m/%d")
print(len(dftiles_large))
print(len(dftiles_large["Key"].unique()))
dftiles_large.head(2)

dftiles_large.columns

## Generating Sinkhole Features

daysdelta = timedelta(0)
fields = ["lon_t", "lat_t", "Key"]
large_tiles_sh = shAttributes(df_sh, dftiles_large, daysdelta, fields)
print(len(large_tiles_sh))
large_tiles_sh.head(2)

large_tiles_sh.to_csv(attrs+"scale_up_F_satellite_ws_soil_karst_sh_attr_F.csv")
